Перенос типов в отдельный файл.
